Refactor `resolve` command to prepare for default component list extraction
