+++ cmd/gardener-landscape-kit/app/app.go
@@ -12,7 +12,7 @@ import (
 
 	"github.com/gardener/gardener-landscape-kit/pkg/cmd"
 	"github.com/gardener/gardener-landscape-kit/pkg/cmd/generate"
-	"github.com/gardener/gardener-landscape-kit/pkg/cmd/resolveocm"
+	"github.com/gardener/gardener-landscape-kit/pkg/cmd/resolve"
 	"github.com/gardener/gardener-landscape-kit/pkg/cmd/version"
 )
 
@@ -46,7 +46,7 @@ func NewCommand() *cobra.Command {
 
 	for _, subcommand := range []*cobra.Command{
 		generate.NewCommand(opts),
-		resolveocm.NewCommand(opts),
+		resolve.NewCommand(opts),
 		version.NewCommand(opts),
 	} {
 		cmd.AddCommand(subcommand)
